refactor: 사다리게임 step2
* player 객체 생성 클래스를 service로 이관함.
