+++ src/main/java/ladderGame/service/CreatePlayer.java
@@ -1,8 +1,9 @@
-package ladderGame.model;
+package ladderGame.service;
 
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
+import ladderGame.model.Player;
 
 public class CreatePlayer {
