+++ src/test/java/ladderGame/service/CreatePlayerTest.java
@@ -1,7 +1,8 @@
-package ladderGame.model;
+package ladderGame.service;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import ladderGame.model.Player;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
